+++ vignettes/articles/precompile.R
@@ -1,0 +1,4 @@
+# Precompiled long-running vignettes
+# Must manually move image files to vignettes/articles/ after knit
+
+knitr::knit("vignettes/articles/covariates.Rmd.orig", output = "vignettes/articles/covariates.Rmd")